Rename _journal to journal
