+++ optuna/storages/journal/_base.py
